requirements.txt
Django==2.2
pytz==2019.2
pymongo==3.11.4
djangorestframework==3.12.4
